refactor(allocator): remove `Arena::alloc_try_with` and `try_alloc_try_with` methods (#21438)
These methods are not useful. Remove them.
